Scripts/ICC: move Sister Svalna to her own file
I've almost guttered her, so lets try it more gently this time
